zk-password-manager/requirements.txt
# Dependencies
cryptography>=41.0.0
argon2-cffi>=23.1.0
flask>=2.3.0
werkzeug>=2.3.0

# Development (optional)
# pytest>=7.4.0
# black>=23.7.0
# mypy>=1.5.0
